# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Smart Home API
  description: API to manage smart home devices
host: 'https://server-o8if.onrender.com'
basePath: '/'
schemes:
  - http
paths:
  /user/{uid}:
    post:
      summary: Update or create user name
      description: Updates a user's name if the UID exists or creates a new user with the given UID if it does not exist.
      parameters:
        - in: path
          name: uid
          required: true
          type: string
          description: The UID of the user to update or create
        - in: body
          name: body
          required: true
          schema:
            required:
              - firstName
              - lastName
            properties:
              firstName:
                type: string
                example: John
              lastName:
                type: string
                example: Doe
      responses:
        200:
          description: User name updated successfully or new user created
        500:
          description: Error updating or creating user
    get:
      summary: Get user names by UID
      description: Retrieves the first and last name of a user by UID.
      parameters:
        - in: path
          name: uid
          required: true
          type: string
          description: The UID of the user
      responses:
        200:
          description: User names retrieved successfully
          schema:
            type: object
            properties:
              firstName:
                type: string
                example: John
              lastName:
                type: string
                example: Doe
        404:
          description: User not found
        500:
          description: Error retrieving user names
  /device/{type}:
    post:
      summary: Update device state
      description: Updates the state of a specific device
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              state:
                type: string
                example: 'on'
      responses:
        200:
          description: Device state updated successfully
        500:
          description: Error updating device state
    get:
      summary: Get device state
      description: Retrieves the current state of a specific device
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
      responses:
        200:
          description: Device state retrieved successfully
        500:
          description: Error retrieving device state
  /device/{type}/state:
    get:
      summary: Read device state
      description: Returns the current state of a specific device
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
      responses:
        200:
          description: Device state retrieved successfully
        500:
          description: Error retrieving device state
  /device/{type}/image:
    get:
      summary: Read device image URL
      description: Returns the image URL of a specific device
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
      responses:
        200:
          description: Device image URL retrieved successfully
        500:
          description: Error retrieving device image URL
  /devices/state:
    get:
      summary: Get All Devices ID, Names and their current state
      description: Retrieves a list of all device IDs, names and their state stored in the database.
      responses:
        200:
          description: A list of device were successfully retrieved.
          schema:
            type: object
            properties:
              deviceNames:
                type: array
                items:
                  type: string
        500:
          description: Error fetching devices.
  /device/{deviceName}/{deviceInformation}:
    post:
      summary: Update specific information of a device
      description: Updates specific information of a device identified by its name.
      parameters:
        - name: deviceName
          in: path
          description: Name of the device to be updated
          required: true
          type: string
        - name: deviceInformation
          in: path
          description: Type of information to be updated
          required: true
          type: string
        - name: newInformation
          in: body
          description: New information to be set for the device
          required: true
          schema:
            type: object
            properties:
              newInformation:
                type: string
      responses:
        200:
          description: Success response
          schema:
            type: object
            properties:
              message:
                type: string
                description: Success message
        500:
          description: Error response
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
  /device/{type}/{password}:
    post:
      summary: Update device state with password
      description: Updates the state of a specific device with password authentication
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
        - in: path
          name: password
          required: true
          type: string
          description: API password for authentication
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              newState:
                type: string
                example: 'on'
      responses:
        200:
          description: Device state updated successfully
        401:
          description: Unauthorized - Invalid password
        500:
          description: Error updating device state
  /device/{type}/{typeInformation}/{password}:
    post:
      summary: Update specific information of a device with password
      description: Updates specific information of a device identified by its type and type information with password authentication
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
        - in: path
          name: typeInformation
          required: true
          type: string
          description: Type of information to be updated
        - in: path
          name: password
          required: true
          type: string
          description: API password for authentication
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              newState:
                type: string
                example: 'newState'
      responses:
        200:
          description: Specific device information updated successfully
        401:
          description: Unauthorized - Invalid password
        500:
          description: Error updating specific device information
  /mediaPlayer/skip:
    post:
      summary: Skip media player channel
      description: Skips the channel on the media player.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              newInformation:
                type: string
                example: 'next'
      responses:
        200:
          description: Skip emitted successfully
        500:
          description: Error skipping media player channel
  /static/device/{type}/{password}:
    post:
      summary: Update device state with password (Static)
      description: Updates the state of a specific device with password authentication (Static)
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
        - in: path
          name: password
          required: true
          type: string
          description: API password for authentication
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              newState:
                type: string
                example: 'on'
      responses:
        200:
          description: Device state updated successfully (Static)
        401:
          description: Unauthorized - Invalid password (Static)
        500:
          description: Error updating device state (Static)
  /static/device/{type}/{typeInformation}/{password}:
    post:
      summary: Update specific information of a device with password (Static)
      description: Updates specific information of a device identified by its type and type information with password authentication (Static)
      parameters:
        - in: path
          name: type
          required: true
          type: string
          description: Type of the device
        - in: path
          name: typeInformation
          required: true
          type: string
          description: Type of information to be updated
        - in: path
          name: password
          required: true
          type: string
          description: API password for authentication
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              newState:
                type: string
                example: 'newState'
      responses:
        200:
          description: Specific device information updated successfully (Static)
        401:
          description: Unauthorized - Invalid password (Static)
        500:
          description: Error updating specific device information (Static)